update: 优化传输系统性能与内存，重构事件与日志对象池，完善集成文档
- **优化传输系统性能**：
  - 在多个资源类（`EnergyResource`、`FluidResource`、`ItemResource`、`ArsSourceResource`、`BotaniaManaResource`、`MekanismChemicalResource`、`MekanismHeatResource`）中新增 `TYPE_ID` 常量并覆写 `isSimpleResource()` 方法，标记能量、魔源、热量等为简单资源，启用快速传输路径。
  - 在 `TransferPipeline` 中为简单资源新增 `transferSimpleResource` 快速路径，跳过模拟提取步骤，直接执行 `extract + insert` 并回退多余量，减少方法调用开销。
  - 在 `CooldownManager` 中将 `HashMap` 替换为 FastUtil 的 `Object2ObjectOpenHashMap` 和 `Object2IntOpenHashMap`，提升集合性能。
  - 在 `FilterEvaluator` 中将编译缓存从 `HashMap` 改为 `WeakHashMap`，允许垃圾回收未使用的缓存条目，减少内存占用。
  - 在 `ContainerConfig` 中将缓存乘数字段类型从 `int` 改为 `long`，避免溢出。
  - 在 `LogisticsTicker` 中优化冷却键计算，使用 `sourceKey ^ (Long.rotateLeft(type.bitOffset(), 48))` 减少冲突。

- **重构事件与日志对象池**：
  - 删除 `BulkExtractionResult` 类，移除批量传输模式，简化传输逻辑。
  - 在 `PreTransferEvent` 和 `PostTransferEvent` 中引入对象池（`ArrayDeque`），新增 `obtain` 和 `recycle` 静态方法，减少事件对象分配开销。
  - 将 `TransferEntry` 从 `record` 改为类并实现对象池，同样新增 `obtain` 和 `recycle` 方法，在 `TransferRecentLog` 中回收旧条目。
  - 在 `TransferPipeline` 中事件发布后调用 `recycle` 回收事件对象。

- **完善集成文档与资源实现**：
  - 更新 `docs/INTEGRATION.md`，补充类型 ID 常量定义、传输上限说明、`canInsertToTarget` 方法示例和完整能量资源示例。
  - 在 `MekanismChemicalResource` 中实现 `canInsertToTarget` 方法，支持化学品传输的目标过滤检查。
  - 在 `BotaniaManaResource` 中调整提取逻辑，即使魔力为 0 也返回句柄，让 `extract` 方法处理；统一 `extract` 和 `insert` 方法的参数与返回类型为 `long`。
  - 在 `NodeConfiguratorScreen` 中调整堆叠上限显示逻辑，使用 `Long.MAX_VALUE / 2` 作为无限判断阈值。

- **优化服务与缓存管理**：
  - 重构 `FaceConfigService`，移除对 `ServerLevel` 和 `DropHandler` 的依赖，在 `get` 方法中延迟设置 `sharedContainerConfig`。
  - 在 `CacheManager` 的 `invalidate` 和 `clear` 方法中设置 `keysDirty = true`，确保键列表及时更新。
  - 在 `FaceConfigHandler` 中移除冗余注释。
  - 在 `ILinkManager` 中移除多余空行，统一注释风格。
  - 在 `LinkManager` 中更新 `FaceConfigService` 构造调用，移除 `level` 和 `dropHandler` 参数。

- **修复与调整**：
  - 在 `BlueprintItem` 中调整恢复计数代码位置，保持逻辑清晰。
  - 在 `NodeStats` 中新增 `firstTransferTime`、`lastTransferTime` 字段和 `getTransfersPerMinute` 方法，支持传输速率计算。
  - 在 `SLJadePlugin` 中将全局传输统计改为面级统计，使用 `NodeStats` 计算速率和上次传输时间，并更新翻译键。
  - 在 `ResourceAdapterProtocol` 中移除批量传输相关方法，新增 `getAdapter` 方法供快速路径使用。
  - 在 `TransferUtils` 中移除 `TransferProtocol` 接口的批量传输默认方法，更新 `SimpleProtocol` 的参数类型为 `Long`。

**行为对比**：
- 之前：传输事件和日志条目每次创建新对象，存在分配开销；简单资源走复杂模拟路径；部分缓存使用强引用；化学品传输缺少目标过滤；集成文档不完整。
- 之后：事件和日志使用对象池，减少 GC 压力；简单资源启用快速路径，提升性能；缓存使用弱引用，内存更友好；化学品支持目标过滤；集成文档提供完整示例和说明。

+++ src/main/java/com/coobird/staticlogistics/storage/model/ContainerConfig.java
@@ -152,9 +152,9 @@ public class ContainerConfig {
             }
         }
 
-        this.cachedSpeedMult = (int) Math.min(speed, INFINITY_MARKER);
-        this.cachedRangeMult = (int) Math.min(range, INFINITY_MARKER);
-        this.cachedStackMult = (int) Math.min(stack, INFINITY_MARKER);
+        this.cachedSpeedMult = Math.min(speed, INFINITY_MARKER);
+        this.cachedRangeMult = Math.min(range, INFINITY_MARKER);
+        this.cachedStackMult = Math.min(stack, INFINITY_MARKER);
         this.cachedDimEffective = dim;
         this.cacheDirty = false;
